Search for all words in a sentence
The Places API returns what thinks is the best match of a Venue. In this sense it doesn't return the exact name as searched but Venues containing the words in the search. By replacing the spaces with the % wildcard we are also looking locally for all entries containing any of the words, being closer to what the API does. This is still helpful for the user as sometimes Venues are named in different ways as what the user could expect. For example Barcelona Restaurant could be named as Restaurant Barcelona.