Update Entity, DAO and Db version to store exact results for a given query (with accepted advantages and drawbacks)
Changed database version to 2

The Foursquare API algorithm returns Venues not only related to the Venue names keyed in. For this reason it is not possible to search locally for the words in the Venue name. Local queries shall be read using the same query. This means that similar queries will not work offline at this stage, but old ones will produce locally the same results than the ones given by the backend.

Cached results are only meant for offline work. Even if they are present and there is a network connection a new call will be made and the response cached for future offline use.